requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest-django
kafka-python
pytest
gunicorn